plumb configOverload better
