+++ shared/desktop/app/kb2-impl.desktop.tsx
@@ -16,7 +16,7 @@ if (pathSep !== '/' && pathSep !== '\\') {
 
 const kb2: KB2['constants'] = {
   assetRoot: path.resolve(__DEV__ ? '.' : app.getAppPath()),
-  configOverload: {},
+  configOverload: {}, // filled in later
   dokanPath: path.resolve(env.LOCALAPPDATA ?? '', 'Keybase', 'DokanSetup_redist.exe'),
   downloadFolder: path.join(os.homedir(), 'Downloads'),
   env: {
