+++ shared/desktop/app/main-window.desktop.tsx
@@ -4,7 +4,7 @@ import * as ConfigGen from '../../actions/config-gen'
 import * as fs from 'fs'
 import menuHelper from './menu-helper.desktop'
 import type {WindowState} from '../../constants/types/config'
-import {showDevTools} from './dynamic-config'
+import {showDevTools} from '../../local-debug'
 import {guiConfigFilename, isDarwin, isWindows, defaultUseNativeFrame} from '../../constants/platform.desktop'
 import logger from '../../logger'
 import debounce from 'lodash/debounce'
