+++ shared/desktop/app/menu-bar.desktop.tsx
@@ -5,7 +5,7 @@ import * as Electron from 'electron'
 import logger from '../../logger'
 import {isDarwin, isWindows, isLinux, getAssetPath} from '../../constants/platform.desktop'
 import {menubar} from 'menubar'
-import {showDevTools, skipSecondaryDevtools} from './dynamic-config'
+import {showDevTools, skipSecondaryDevtools} from '../../local-debug'
 import {getMainWindow} from './main-window.desktop'
 import getIcons from '../../menubar/icons'
 import os from 'os'
