+++ shared/local-debug.d.ts
@@ -1,3 +1,4 @@
+export declare const allowMultipleInstances: boolean
 export declare const enableActionLogging: boolean
 export declare const enableStoreLogging: boolean
 export declare const featureFlagsOverride: boolean
@@ -12,8 +13,10 @@ export declare const printOutstandingRPCs: boolean
 export declare const printOutstandingTimerListeners: boolean
 export declare const printRPC: boolean
 export declare const printRPCBytes: boolean
-export declare const printRPCWaitingSession: boolean
 export declare const printRPCStats: boolean
+export declare const printRPCWaitingSession: boolean
+export declare const showDevTools: boolean
 export declare const skipAppFocusActions: boolean
+export declare const skipSecondaryDevtools: boolean
 export declare const userTimings: boolean
 export declare const virtualListMarks: boolean
